CharactersTable tbody styling
Changed table styling with a combination of CSS and JS. Set tbody display as block,
thus allowing it to have a set height. Since this breaks the width of tbody and td's,
these are set via JS in useEffect hooks. Additionally, attached an onresize
event listener which (re)styles the table.